feat: add scroll menu event, change gallery event
