+++ index.html
@@ -23,8 +23,7 @@
                 <a href="#contact">contact</a>
             </nav>
         </header>
-        <section class="slider">
-            <a name="home"></a>
+        <section class="slider" id="home">
             <div class="slider__content">
                  <div class="slider__item slider__item_active">
                     <div class="item__phone">
@@ -56,8 +55,7 @@
                 </div>
             </div>
         </section>
-        <section class="services">
-            <a name="services"></a>
+        <section class="services" id="services">
             <div class="services__title">Our Services</div>
             <div class="services__description">
                 Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum. Duis mollis, non commodo luctus, 
@@ -144,8 +142,7 @@
                 </div>
             </div>
         </section>
-        <section class="portfolio">
-            <a name="portfolio"></a>
+        <section class="portfolio" id="portfolio">
             <div class="portfolio__title">Portfolio</div>
             <div class="portfolio__description">Duis mollis, est non commodo luctus, nisi erat porttitor ligula, eget lacinia odio sem nec elit.</div>
             <div class="portfolio__filter">
@@ -193,8 +190,7 @@
                 </div>
             </div>
         </section>
-        <section class="about">
-            <a name="about"></a>
+        <section class="about" id="about">
             <div class="about__title">About Us</div>
             <div class="about__description">
                 Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum.
@@ -260,8 +256,7 @@
                 </div>
             </div>
         </section>
-        <section class="contact">
-            <a name="contact"></a>
+        <section class="contact" id="contact">
             <div class="contact__title">Get a contact</div>
             <div class="contact__description">
                 Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum.
@@ -330,5 +325,5 @@
         </div>
     </div>
 </body>
-<script src="script.js"></script>
+<script src="script.mjs" type="module"></script>
 </html>
